Remove interface pages.
Refs #8579
