+++ src/gui/app/source/styles/Network.less
@@ -6,8 +6,6 @@
 @network-bg-color: #f5f5f5;
 
 .network-overview {
-  border-top  : 1px solid @network-border-color;
-  margin-top  : 20px;
   padding-top : 20px;
 
   .section {
